requirements.txt
attrs==18.2.0
lxml==4.2.5
sortedcontainers==2.0.5